% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vowels.R
\name{basic_vclass}
\alias{basic_vclass}
\title{basic vclass coding
Input must be output of `word_phone_nest`}
\usage{
basic_vclass(df)
}
\description{
basic vclass coding
Input must be output of `word_phone_nest`
}
