% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texgrid_processing.R
\name{enrich_syl}
\alias{enrich_syl}
\title{enrich with syllable structure
must be output of `word_phone_nest`}
\usage{
enrich_syl(df)
}
\description{
enrich with syllable structure
must be output of `word_phone_nest`
}
