% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phila_system.R
\name{phila_lowback}
\alias{phila_lowback}
\title{marry fixing}
\usage{
phila_lowback(df)
}
\description{
input expects basic vclass coding
}
\details{
todo: check for vclass column. Throw error otherwise
}
