% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phila_system.R
\name{phila_marry}
\alias{phila_marry}
\title{marry fixing}
\usage{
phila_marry(df)
}
\description{
input expects basic vclass coding
}
\details{
todo: check for vclass column. Throw error otherwise

Dramatically incomplete!
}
